+++ apple/TRssReader/ViewControllers/LoginViewController.swift
@@ -1,0 +1,21 @@
+//
+//  LoginViewController.swift
+//  TRssReader
+//
+//  Created by Ty Hopp on 13/4/23.
+//
+
+import SwiftUI
+
+// TODO: Implement
+struct LoginViewController: View {
+    var body: some View {
+        Text("Login view controller")
+    }
+}
+
+struct LoginViewController_Previews: PreviewProvider {
+    static var previews: some View {
+        LoginViewController()
+    }
+}
